[vulkan] Adjusted pipeline keys inside patchControlPoints

+++ src/video_core/renderer_vulkan/vk_graphics_pipeline.cpp
@@ -677,7 +677,7 @@ void GraphicsPipeline::MakePipeline(VkRenderPass render_pass) {
         .sType = VK_STRUCTURE_TYPE_PIPELINE_TESSELLATION_STATE_CREATE_INFO,
         .pNext = nullptr,
         .flags = 0,
-        .patchControlPoints = key.state.patch_control_points_minus_one.Value() + 1,
+        .patchControlPoints = static_cast<u32>(key.state.patch_control_points_minus_one.Value()) + 1,
     };
     std::array<VkViewportSwizzleNV, Maxwell::NumViewports> swizzles;
     std::ranges::transform(key.state.viewport_swizzles, swizzles.begin(), UnpackViewportSwizzle);
